requirements.txt
Flask
gunicorn
flask-sqlalchemy
requests
